Bug 1331498: Update libvpx to v1.6.1. r=rillian
Removes armv6 and mmx optimizations.

Removes x86inc.asm distinction. This was put in place because of issues
with 32bit PIC builds but x86inc.asm has since been improved to support
those.

Switch to '.S' for arm assembly. There is now an #include for a helper
file. All build systems appear to support .S so switch to it for
consistency.

Remove clang-cl.patch. There have been numerous changes upstream to
improve clang support.

MozReview-Commit-ID: IHVTbqSY2U7

+++ media/libvpx/libvpx/vpx_dsp/x86/variance_avx2.c
@@ -14,13 +14,13 @@ typedef void (*get_var_avx2)(const uint8_t *src, int src_stride,
                              unsigned int *sse, int *sum);
 
 void vpx_get32x32var_avx2(const uint8_t *src, int src_stride,
-                          const uint8_t *ref, int ref_stride,
-                          unsigned int *sse, int *sum);
+                          const uint8_t *ref, int ref_stride, unsigned int *sse,
+                          int *sum);
 
 static void variance_avx2(const uint8_t *src, int src_stride,
-                          const uint8_t *ref, int  ref_stride,
-                          int w, int h, unsigned int *sse, int *sum,
-                          get_var_avx2 var_fn, int block_size) {
+                          const uint8_t *ref, int ref_stride, int w, int h,
+                          unsigned int *sse, int *sum, get_var_avx2 var_fn,
+                          int block_size) {
   int i, j;
 
   *sse = 0;
@@ -30,21 +30,20 @@ static void variance_avx2(const uint8_t *src, int src_stride,
     for (j = 0; j < w; j += block_size) {
       unsigned int sse0;
       int sum0;
-      var_fn(&src[src_stride * i + j], src_stride,
-             &ref[ref_stride * i + j], ref_stride, &sse0, &sum0);
+      var_fn(&src[src_stride * i + j], src_stride, &ref[ref_stride * i + j],
+             ref_stride, &sse0, &sum0);
       *sse += sse0;
       *sum += sum0;
     }
   }
 }
 
-
 unsigned int vpx_variance16x16_avx2(const uint8_t *src, int src_stride,
                                     const uint8_t *ref, int ref_stride,
                                     unsigned int *sse) {
   int sum;
-  variance_avx2(src, src_stride, ref, ref_stride, 16, 16,
-                sse, &sum, vpx_get16x16var_avx2, 16);
+  variance_avx2(src, src_stride, ref, ref_stride, 16, 16, sse, &sum,
+                vpx_get16x16var_avx2, 16);
   return *sse - (((uint32_t)((int64_t)sum * sum)) >> 8);
 }
 
@@ -60,124 +59,97 @@ unsigned int vpx_variance32x16_avx2(const uint8_t *src, int src_stride,
                                     const uint8_t *ref, int ref_stride,
                                     unsigned int *sse) {
   int sum;
-  variance_avx2(src, src_stride, ref, ref_stride, 32, 16,
-                sse, &sum, vpx_get32x32var_avx2, 32);
-  return *sse - (((int64_t)sum * sum) >> 9);
+  variance_avx2(src, src_stride, ref, ref_stride, 32, 16, sse, &sum,
+                vpx_get32x32var_avx2, 32);
+  return *sse - (uint32_t)(((int64_t)sum * sum) >> 9);
 }
 
 unsigned int vpx_variance32x32_avx2(const uint8_t *src, int src_stride,
                                     const uint8_t *ref, int ref_stride,
                                     unsigned int *sse) {
   int sum;
-  variance_avx2(src, src_stride, ref, ref_stride, 32, 32,
-                sse, &sum, vpx_get32x32var_avx2, 32);
-  return *sse - (((int64_t)sum * sum) >> 10);
+  variance_avx2(src, src_stride, ref, ref_stride, 32, 32, sse, &sum,
+                vpx_get32x32var_avx2, 32);
+  return *sse - (uint32_t)(((int64_t)sum * sum) >> 10);
 }
 
 unsigned int vpx_variance64x64_avx2(const uint8_t *src, int src_stride,
                                     const uint8_t *ref, int ref_stride,
                                     unsigned int *sse) {
   int sum;
-  variance_avx2(src, src_stride, ref, ref_stride, 64, 64,
-                sse, &sum, vpx_get32x32var_avx2, 32);
-  return *sse - (((int64_t)sum * sum) >> 12);
+  variance_avx2(src, src_stride, ref, ref_stride, 64, 64, sse, &sum,
+                vpx_get32x32var_avx2, 32);
+  return *sse - (uint32_t)(((int64_t)sum * sum) >> 12);
 }
 
 unsigned int vpx_variance64x32_avx2(const uint8_t *src, int src_stride,
                                     const uint8_t *ref, int ref_stride,
                                     unsigned int *sse) {
   int sum;
-  variance_avx2(src, src_stride, ref, ref_stride, 64, 32,
-                sse, &sum, vpx_get32x32var_avx2, 32);
-  return *sse - (((int64_t)sum * sum) >> 11);
+  variance_avx2(src, src_stride, ref, ref_stride, 64, 32, sse, &sum,
+                vpx_get32x32var_avx2, 32);
+  return *sse - (uint32_t)(((int64_t)sum * sum) >> 11);
 }
 
 unsigned int vpx_sub_pixel_variance32xh_avx2(const uint8_t *src, int src_stride,
                                              int x_offset, int y_offset,
                                              const uint8_t *dst, int dst_stride,
-                                             int height,
-                                             unsigned int *sse);
+                                             int height, unsigned int *sse);
 
-unsigned int vpx_sub_pixel_avg_variance32xh_avx2(const uint8_t *src,
-                                                 int src_stride,
-                                                 int x_offset,
-                                                 int y_offset,
-                                                 const uint8_t *dst,
-                                                 int dst_stride,
-                                                 const uint8_t *sec,
-                                                 int sec_stride,
-                                                 int height,
-                                                 unsigned int *sseptr);
+unsigned int vpx_sub_pixel_avg_variance32xh_avx2(
+    const uint8_t *src, int src_stride, int x_offset, int y_offset,
+    const uint8_t *dst, int dst_stride, const uint8_t *sec, int sec_stride,
+    int height, unsigned int *sseptr);
 
 unsigned int vpx_sub_pixel_variance64x64_avx2(const uint8_t *src,
-                                              int src_stride,
-                                              int x_offset,
-                                              int y_offset,
-                                              const uint8_t *dst,
+                                              int src_stride, int x_offset,
+                                              int y_offset, const uint8_t *dst,
                                               int dst_stride,
                                               unsigned int *sse) {
   unsigned int sse1;
-  const int se1 = vpx_sub_pixel_variance32xh_avx2(src, src_stride, x_offset,
-                                                  y_offset, dst, dst_stride,
-                                                  64, &sse1);
+  const int se1 = vpx_sub_pixel_variance32xh_avx2(
+      src, src_stride, x_offset, y_offset, dst, dst_stride, 64, &sse1);
   unsigned int sse2;
-  const int se2 = vpx_sub_pixel_variance32xh_avx2(src + 32, src_stride,
-                                                  x_offset, y_offset,
-                                                  dst + 32, dst_stride,
-                                                  64, &sse2);
+  const int se2 =
+      vpx_sub_pixel_variance32xh_avx2(src + 32, src_stride, x_offset, y_offset,
+                                      dst + 32, dst_stride, 64, &sse2);
   const int se = se1 + se2;
   *sse = sse1 + sse2;
-  return *sse - (((int64_t)se * se) >> 12);
+  return *sse - (uint32_t)(((int64_t)se * se) >> 12);
 }
 
 unsigned int vpx_sub_pixel_variance32x32_avx2(const uint8_t *src,
-                                              int src_stride,
-                                              int x_offset,
-                                              int y_offset,
-                                              const uint8_t *dst,
+                                              int src_stride, int x_offset,
+                                              int y_offset, const uint8_t *dst,
                                               int dst_stride,
                                               unsigned int *sse) {
-  const int se = vpx_sub_pixel_variance32xh_avx2(src, src_stride, x_offset,
-                                                 y_offset, dst, dst_stride,
-                                                 32, sse);
-  return *sse - (((int64_t)se * se) >> 10);
+  const int se = vpx_sub_pixel_variance32xh_avx2(
+      src, src_stride, x_offset, y_offset, dst, dst_stride, 32, sse);
+  return *sse - (uint32_t)(((int64_t)se * se) >> 10);
 }
 
-unsigned int vpx_sub_pixel_avg_variance64x64_avx2(const uint8_t *src,
-                                                  int src_stride,
-                                                  int x_offset,
-                                                  int y_offset,
-                                                  const uint8_t *dst,
-                                                  int dst_stride,
-                                                  unsigned int *sse,
-                                                  const uint8_t *sec) {
+unsigned int vpx_sub_pixel_avg_variance64x64_avx2(
+    const uint8_t *src, int src_stride, int x_offset, int y_offset,
+    const uint8_t *dst, int dst_stride, unsigned int *sse, const uint8_t *sec) {
   unsigned int sse1;
-  const int se1 = vpx_sub_pixel_avg_variance32xh_avx2(src, src_stride, x_offset,
-                                                      y_offset, dst, dst_stride,
-                                                      sec, 64, 64, &sse1);
+  const int se1 = vpx_sub_pixel_avg_variance32xh_avx2(
+      src, src_stride, x_offset, y_offset, dst, dst_stride, sec, 64, 64, &sse1);
   unsigned int sse2;
-  const int se2 =
-    vpx_sub_pixel_avg_variance32xh_avx2(src + 32, src_stride, x_offset,
-                                        y_offset, dst + 32, dst_stride,
-                                        sec + 32, 64, 64, &sse2);
+  const int se2 = vpx_sub_pixel_avg_variance32xh_avx2(
+      src + 32, src_stride, x_offset, y_offset, dst + 32, dst_stride, sec + 32,
+      64, 64, &sse2);
   const int se = se1 + se2;
 
   *sse = sse1 + sse2;
 
-  return *sse - (((int64_t)se * se) >> 12);
+  return *sse - (uint32_t)(((int64_t)se * se) >> 12);
 }
 
-unsigned int vpx_sub_pixel_avg_variance32x32_avx2(const uint8_t *src,
-                                                  int src_stride,
-                                                  int x_offset,
-                                                  int y_offset,
-                                                  const uint8_t *dst,
-                                                  int dst_stride,
-                                                  unsigned int *sse,
-                                                  const uint8_t *sec) {
+unsigned int vpx_sub_pixel_avg_variance32x32_avx2(
+    const uint8_t *src, int src_stride, int x_offset, int y_offset,
+    const uint8_t *dst, int dst_stride, unsigned int *sse, const uint8_t *sec) {
   // Process 32 elements in parallel.
-  const int se = vpx_sub_pixel_avg_variance32xh_avx2(src, src_stride, x_offset,
-                                                     y_offset, dst, dst_stride,
-                                                     sec, 32, 32, sse);
-  return *sse - (((int64_t)se * se) >> 10);
+  const int se = vpx_sub_pixel_avg_variance32xh_avx2(
+      src, src_stride, x_offset, y_offset, dst, dst_stride, sec, 32, 32, sse);
+  return *sse - (uint32_t)(((int64_t)se * se) >> 10);
 }
